+++ Flow/src/utopia/flow/view/immutable/View.scala
@@ -59,7 +59,7 @@ object View
   * @since 4.11.2020, v1.9
   * @tparam A Type of value viewable though this wrapper
   */
-trait View[+A] extends Any
+trait View[+A]
 {
 	// ABSTRACT	----------------------------